extract filters component to remove uneccessary state causing stats page to re-render
